robustvsnormalzoo: fixing loading of models
The uploaded models were trained with PTL which serialized more than the state_dict. To unserialize we need a dummy utils.py with any MyCheckpoint class on PATH. This ugly fix adds the script path to the sys path before loading and removes it after. May interfere if user has other utils.py